+++ packages/amplify-cli/src/commands/status.ts
@@ -1,15 +1,6 @@
-import * as path from 'path';
-import {
-  ViewResourceTableParams,
-  CLIParams,
-  $TSAny,
-  $TSContext,
-  pathManager,
-  stateManager,
-  ApiCategoryFacade,
-} from '@aws-amplify/amplify-cli-core';
+import { ViewResourceTableParams, CLIParams, $TSContext } from '@aws-amplify/amplify-cli-core';
 import { printer } from '@aws-amplify/amplify-prompts';
-import { readProjectSchema } from 'graphql-transformer-core';
+import { showApiAuthAcm } from '@aws-amplify/amplify-category-api';
 
 /**
  * Entry point for status command
@@ -33,7 +24,7 @@ export const run = async (context: $TSContext): Promise<void> => {
         return;
       }
 
-      await showApiAuthAcm(context);
+      await showApiAuthAcm(context, cliParams.cliOptions.acm);
     } catch (err) {
       printer.error(err?.message);
     }
@@ -53,58 +44,4 @@ const showAmplifyConsoleHostingStatus = async (context: $TSContext): Promise<voi
       await status(context);
     }
   }
-};
-
-const showApiAuthAcm = async (context): Promise<void> => {
-  const providerPlugin = await import(context.amplify.getProviderPlugins(context)?.awscloudformation);
-  const transformerVersion = await ApiCategoryFacade.getTransformerVersion(context);
-
-  if (transformerVersion < 2) {
-    printer.error('This command requires version two or greater of the GraphQL transformer.');
-    return;
-  }
-
-  const apiNames = Object.entries(stateManager.getMeta()?.api || {})
-    .filter(([, apiResource]) => (apiResource as $TSAny).service === 'AppSync')
-    .map(([name]) => name);
-
-  if (apiNames.length === 0) {
-    printer.info(
-      'No GraphQL API configured in the project. Only GraphQL APIs can be migrated. To add a GraphQL API run `amplify add api`.',
-    );
-    return;
-  }
-
-  if (apiNames.length > 1) {
-    // this condition should never hit as we only allow a single GraphQL API per project.
-    printer.error(
-      'You have multiple GraphQL APIs in the project. Only one GraphQL API is allowed per project. Run `amplify remove api` to remove an API.',
-    );
-    return;
-  }
-
-  // Do a full schema compilation to make sure we are not printing an ACM for an invalid schema
-  try {
-    await providerPlugin.compileSchema(context, {
-      forceCompile: true,
-    });
-  } catch (error) {
-    printer.warn('ACM generation requires a valid schema, the provided schema is invalid.');
-
-    if (error.name) {
-      printer.error(`${error.name}: ${error.message?.trim()}`);
-    } else {
-      printer.error(`An error has occurred during schema compilation: ${error.message?.trim()}`);
-    }
-
-    return;
-  }
-
-  const apiName = apiNames[0];
-  const apiResourceDir = path.join(pathManager.getBackendDirPath(), 'api', apiName);
-  const schema = await readProjectSchema(apiResourceDir);
-  const cliOptions = context?.input?.options ?? {};
-  const { showACM } = await import('../extensions/amplify-helpers/show-auth-acm');
-
-  showACM(schema, cliOptions.acm);
 };