+++ packages/amplify-provider-awscloudformation/src/index.ts
@@ -54,7 +54,6 @@ import { prePushCfnTemplateModifier } from './pre-push-cfn-processor/pre-push-cf
 import { getApiKeyConfig } from './utils/api-key-helpers';
 import { deleteEnvironmentParametersFromService } from './utils/ssm-utils/delete-ssm-parameters';
 export { deleteEnvironmentParametersFromService } from './utils/ssm-utils/delete-ssm-parameters';
-
 import { getEnvParametersUploadHandler, getEnvParametersDownloadHandler } from './utils/ssm-utils/env-parameter-ssm-helpers';
 export {
   getEnvParametersUploadHandler,