+++ packages/amplify-console-hosting/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.8](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-console-hosting@2.5.6...@aws-amplify/amplify-console-hosting@2.5.8) (2023-05-25)
+
+
+### Bug Fixes
+
+* version bump ([f820824](https://github.com/aws-amplify/amplify-cli/commit/f82082416187bd1dd33de9f7b35753026ac17eea))
+
+
+
+
+
+## [2.5.7](https://github.com/sdstolworthy/amplify-cli/compare/@aws-amplify/amplify-console-hosting@2.5.5...@aws-amplify/amplify-console-hosting@2.5.7) (2023-05-25)
+
+**Note:** Version bump only for package @aws-amplify/amplify-console-hosting
+
+
+
+
+
+## [2.5.6](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-console-hosting@2.5.5...@aws-amplify/amplify-console-hosting@2.5.6) (2023-05-25)
+
+**Note:** Version bump only for package @aws-amplify/amplify-console-hosting
+
+
+
+
+
 ## [2.5.5](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-console-hosting@2.5.4...@aws-amplify/amplify-console-hosting@2.5.5) (2023-05-11)